@@ -1,32 +1,23 @@
-﻿using System;
-using Domain.Validators;
+﻿using System.ComponentModel.DataAnnotations;
 
 namespace Domain.Entities
-
 {
-    public class Topic : Entity
+    public class Topic 
     {
-        //private DateTime? createdAt;
+        public int Id { get; set; }
 
+        [Required]
         public string Name{ get; private set; }
-        public Guid CategoryId { get; private set; }
-        public DateTime CreatedAt { get;  set; }
-        //public DateTime? CreatedAt
-        //{
-        //    get { return createdAt; }
-        //    set { createdAt = value == null ? DateTime.UtcNow : value; }
-        //}
-        public DateTime UpdatedAt { get; set; }
-        public Topic(Guid id,string name, Guid categoryId)
+        [Required]
+        public int CategoryId { get; private set; }
+        
+        public Topic(int id, string name, int categoryId)
         {
             Id = id;
             Name = name;
             CategoryId = categoryId;
-            Validate(this,new TopicValidator());
         }
-        public Topic()
-        {
 
-        }
+
     }
 }
