@@ -1,42 +1,27 @@
 ﻿using Infrastructure.Data.Entity.Entities;
 using Infrastructure.Data.Entity.Map;
 using Microsoft.EntityFrameworkCore;
-using System;
-using System.Collections.Generic;
 using System.ComponentModel.DataAnnotations;
 
 namespace Infrastructure.Context
 {
     public class ApplicationContext : DbContext
     {
+        public ApplicationContext(DbContextOptions<ApplicationContext> options)
+        : base(options) { }
+
+
         public DbSet<Category> Category { get; set; }
         public DbSet<Topic> Topic { get; set; }
         public DbSet<Post> Post { get; set; }
-        public DbSet<User> User { get; set; }
         public DbSet<Comment> Comment { get; set; }
 
-        protected override void OnConfiguring(DbContextOptionsBuilder optionsBuilder)
-        {
-            if (Environment.GetEnvironmentVariable("DATABASE_CONN") != null)
-            {
-                optionsBuilder.UseNpgsql(Environment.GetEnvironmentVariable("DATABASE_CONN"), npgsqlOptionsAction: options =>
-                {
-                    options.EnableRetryOnFailure(2, TimeSpan.FromSeconds(5), new List<string>());
-                    options.MigrationsHistoryTable("_MigrationHistory", "webforum");
-                });
-            }
-            else
-            {
-                optionsBuilder.UseInMemoryDatabase("InMemoryProvider");
-            }
-        }
     
         protected override void OnModelCreating(ModelBuilder modelBuilder)
         {
             modelBuilder.ApplyConfiguration(new CategoryMap());
             modelBuilder.ApplyConfiguration(new TopicMap());
             modelBuilder.ApplyConfiguration(new PostMap());
-            modelBuilder.ApplyConfiguration(new UserMap());
             modelBuilder.ApplyConfiguration(new CommentMap());
             modelBuilder.Ignore<ValidationResult>();
 
