@@ -4,7 +4,7 @@ namespace Infrastructure.Data.Entity.Entities
 {
     public class Category
     {
-        public Guid Id { get; set; }
+        public int Id { get; set; }
         public string Name { get; set; }
         public DateTime? CreatedAt { get; set; }
         public DateTime? UpdatedAt { get; set; }
