@@ -4,11 +4,9 @@ namespace Infrastructure.Data.Entity.Entities
 {
     public class Comment
     {
-        public Guid Id { get; set; }
-        public Guid PostId { get; set; }
-        public Guid UserId { get; set; }
+        public int Id { get; set; }
+        public int PostId { get; set; }
         public Post Post { get; set; }
-        public User User { get; set; }
         public string Content { get; set; }
         public DateTime CreatedAt { get; set; }
         public DateTime UpdatedAt { get; set; }
