@@ -4,8 +4,8 @@ namespace Infrastructure.Data.Entity.Entities
 {
     public class Topic
     {
-        public Guid Id { get; set; }
-        public Guid CategoryId { get; set; }
+        public int Id { get; set; }
+        public int CategoryId { get; set; }
         public Category Category { get; set; }
         public Post Post{ get; set; }
         public string Name { get; set; }
