@@ -11,7 +11,17 @@ namespace Infrastructure.Data.Entity.Map
         {
             builder.ToTable("Category", "webforum");
             builder.HasKey(k => k.Id);
-            builder.Property(p => p.Name).IsRequired();
+            builder.Property(p => p.Id)
+            .ValueGeneratedOnAdd();
+
+            builder.Property(p => p.Name).IsRequired().HasMaxLength(255);
+            builder.Property(p => p.CreatedAt)
+            .HasDefaultValueSql("CURRENT_TIMESTAMP");
+
+            builder.Property(p => p.UpdatedAt)
+            .HasDefaultValueSql("CURRENT_TIMESTAMP")
+            .ValueGeneratedOnAddOrUpdate();
+
         }
     }
 }
