@@ -10,9 +10,16 @@ namespace Infrastructure.Data.Entity.Map
         {
             builder.ToTable("Comment", "webforum");
             builder.HasKey(k => k.Id);
+            builder.Property(p => p.Id)
+           .ValueGeneratedOnAdd();
             builder.HasOne(f => f.Post).WithOne(x => x.Comment).HasForeignKey<Comment>(x => x.PostId);
-            builder.HasOne(f => f.User).WithOne(x => x.Comment).HasForeignKey<Comment>(x => x.UserId);
             builder.Property(x => x.Content).IsRequired();
+            builder.Property(p => p.CreatedAt)
+            .HasDefaultValueSql("CURRENT_TIMESTAMP");
+
+            builder.Property(p => p.UpdatedAt)
+            .HasDefaultValueSql("CURRENT_TIMESTAMP")
+            .ValueGeneratedOnAddOrUpdate();
         }
     }
 }
